requirements.txt
pydriller>=2.9
pandas
tqdm
matplotlib
radon  # Python code complexity metrics
python-dotenv
requests
